feat(backend): Various updates including standardization of param variables, logic debugging and clarification, and refactoring to omit obsolete imports.

+++ server/src/controllers/userController.ts
@@ -1,11 +1,10 @@
-import { Request, Response, NextFunction } from 'express';
-import { PrismaClient } from '@prisma/client';
+import { Request, Response } from 'express';
+import { PrismaClient, GlobalRole } from '@prisma/client';
 import { hashPassword } from '../utilities/password';
 import { buildLogEvent } from '../services/buildLogEvent';
 import { generateDiff } from '../services/generateDiff';
 import { getStorageType } from '../config/storage';
 import { storageDispatcher } from '../utilities/storageDispatcher';
-import { CustomRequest } from '../types/CustomRequest';
 import { deleteUserCascade } from '../services/deletionServices/deleteUserCascade';
 
 // Get all users
@@ -16,66 +15,75 @@ export async function getAllUsers(
 ) {
   try {
     const users = await prisma.user.findMany();
-    res.status(200).json(users);
+    res.status(200).json({ data: users });
+    return;
   } catch (error) {
     console.error('Error fetching users: ', error);
     res.status(500).json({ error: 'Failed to fetch users' });
+    return;
   }
 }
 
-// Get user by Id
+// Get user by Id and email
 export async function getUser(
   req: Request,
   res: Response,
   prisma: PrismaClient
 ) {
   try {
-    const { id, email } = req.query;
+    const { userId, userEmail } = req.params;
 
-    if (!id && !email) {
+    if (!userId && !userEmail) {
       return res.status(400).json({ error: 'User ID or email is required' });
     }
 
     let user;
 
-    if (id) {
-      const userId = parseInt(String(id), 10);
-      if (isNaN(userId)) {
-        return res.status(400).json({ error: 'Invalid user ID' });
+    if (userId) {
+      const userIdParsed = parseInt(userId, 10);
+      if (isNaN(userIdParsed)) {
+        res.status(400).json({ error: 'Invalid user ID' });
+        return;
       }
       // Fetch user by ID
-      user = await prisma.user.findUnique({ where: { id: userId } });
-    } else if (email) {
+      user = await prisma.user.findUnique({ where: { id: userIdParsed } });
+    } else if (userEmail) {
       user = await prisma.user.findUnique({
-        where: { email: typeof email === 'string' ? email : undefined },
+        where: { email: typeof userEmail === 'string' ? userEmail : undefined },
       });
     }
 
-    if (!user || user.deletedAt) {
+    if (!user || user.isDeleted) {
       return res.status(404).json({ error: 'User not found' });
     }
 
-    res.status(200).json(user);
+    res.status(200).json({
+      message: 'User was retrieved successfully by user Id',
+      data: user,
+    });
+    return;
   } catch (error) {
     console.error('Error fetching users: ', error);
     res.status(500).json({ error: 'Failed to fetch user' });
+    return;
   }
 }
 
-// Get user by ProjectId
+// Get users by ProjectId
 export async function getUserByProjectId(
   req: Request,
   res: Response,
   prisma: PrismaClient
 ) {
   try {
-    const { userId } = req.params;
-    const projectIdNumber = parseInt(userId, 10);
-    if (isNaN(projectIdNumber)) {
-      return res.status(400).json({ error: 'Invalid project ID' });
+    const { projectId } = req.params;
+    const projectIdParsed = parseInt(projectId, 10);
+    if (isNaN(projectIdParsed)) {
+      res.status(400).json({ error: 'Invalid project ID' });
+      return;
     }
     const users = await prisma.projectMember.findMany({
-      where: { projectId: projectIdNumber },
+      where: { projectId: projectIdParsed },
       include: {
         user: {
           select: {
@@ -91,10 +99,15 @@ export async function getUserByProjectId(
     if (!users || users.length === 0) {
       return res.status(404).json({ error: 'No users found for this project' });
     }
-    res.status(200).json(users.map((member) => member.user));
+    res.status(200).json({
+      message: 'Users successfully fetched by project Id',
+      data: users.map((member) => member.user),
+    });
+    return;
   } catch (error) {
     console.error('Error fetching users by project ID: ', error);
     res.status(500).json({ error: 'Failed to fetch users by project ID' });
+    return;
   }
 }
 
@@ -102,12 +115,11 @@ export async function getUserByProjectId(
 export async function createUser(
   req: Request,
   res: Response,
-  next: NextFunction,
   prisma: PrismaClient
 ) {
   try {
-    const { email, firstName, lastName, passwordHash, role } = req.body;
-    const hashedPassword = await hashPassword(passwordHash);
+    const { email, firstName, lastName, password, role } = req.body;
+    const hashedPassword = await hashPassword(password);
 
     const user = await prisma.user.create({
       data: {
@@ -131,50 +143,54 @@ export async function createUser(
         email: `${user.email}`,
       },
     });
-    res.status(201).json(user);
+    res.status(201).json({ message: 'User successfully created.', data: user });
+    return;
   } catch (error) {
     console.error('Error creating user: ', error);
     res.status(500).json({ error: 'Failed to create user' });
+    return;
   }
 }
 
 // Update user
 export async function updateUser(
-  req: CustomRequest,
+  req: Request,
   res: Response,
-  next: NextFunction,
   prisma: PrismaClient
 ) {
+  // Information about the user performing this action
+  const userInfo: { id: number; globalRole: GlobalRole } = res.locals.userInfo;
+
+  // Pertains to userId payload sent to endpoint
+  const { userId } = req.params;
+  const userIdParsed = parseInt(userId, 10);
+
+  // Check if a user is trying to change their own role
+  const isAdmin =
+    userInfo.globalRole === GlobalRole.ADMIN ||
+    userInfo.globalRole === GlobalRole.SUPERADMIN;
+  const isSelf = userInfo.id === userIdParsed;
+
+  const userData = req.body;
+
+  if (!isAdmin || isSelf)
+    if (userData?.globalRole) {
+      return res.status(403).json({
+        error: 'Unauthorized to change your own global role.',
+      });
+    }
+
   try {
-    const userData = req.body;
-    const { id } = req.params;
-    const userId = parseInt(id, 10);
-
-    if (!req.user) {
-      return res.status(403).json({ error: 'Unauthorized' });
-    }
-
-    // Check if a user is trying to change their own role
-    if ('globalRole' in req.body) {
-      const isAdmin = req.user?.globalRole === 'ADMIN';
-      const isSelf = req.user?.id === userId;
-
-      if (!isAdmin || isSelf) {
-        return res.status(403).json({
-          error: 'Unauthorized to change your own global role.',
-        });
-      }
-    }
-
     const oldUser = await prisma.user.findUnique({
-      where: { id: userId },
+      where: { id: userIdParsed },
     });
     if (!oldUser) {
-      return res.status(404).json({ error: 'User not found' });
+      res.status(404).json({ error: 'User not found' });
+      return;
     }
 
     const newUser = await prisma.user.update({
-      where: { id: userId },
+      where: { id: userIdParsed },
       data: {
         ...userData,
       },
@@ -183,25 +199,23 @@ export async function updateUser(
     const changes = generateDiff(oldUser, newUser);
 
     res.locals.logEvent = buildLogEvent({
-      userId: userId,
+      userId: userIdParsed,
       actorType: 'USER',
       action: 'UPDATE_USER',
-      targetId: userId,
+      targetId: userIdParsed,
       targetType: 'USER',
       metadata: {
         name: `${newUser.firstName}_${newUser.lastName}`,
         changes,
       },
-      ticketId: null,
-      boardId: null,
-      projectId: null,
     });
 
     res.status(200).json(newUser);
-    next();
+    return;
   } catch (error) {
     console.error('Error editing user: ', error);
     res.status(500).json({ error: 'Failed to edit user' });
+    return;
   }
 }
 
@@ -209,7 +223,6 @@ export async function updateUser(
 export async function deleteUser(
   req: Request,
   res: Response,
-  next: NextFunction,
   prisma: PrismaClient
 ) {
   try {
@@ -220,10 +233,11 @@ export async function deleteUser(
       where: { id: userIdParsed },
     });
 
-    if (!existingUser) {
-      return res
+    if (!existingUser || existingUser.isDeleted === true) {
+      res
         .status(404)
         .json({ error: `User with ID ${userIdParsed} not found.` });
+      return;
     }
 
     await deleteUserCascade(userIdParsed);
@@ -243,15 +257,11 @@ export async function deleteUser(
         name: `${deletedUserData.firstName}_${deletedUserData.lastName}`,
         deletedOn: `${deletedUserData.deletedAt}`,
       },
-      ticketId: null,
-      boardId: null,
-      projectId: null,
     });
 
     res.status(200).json({
       message: `Soft delete action successful. Date ${deletedUserData.deletedAt} added to deletedAt`,
     });
-    next();
   } catch (error) {
     console.error('Error soft-deleting user: ', error);
     res.status(500).json({ error: 'Failed to soft-delete user' });
@@ -265,8 +275,8 @@ export async function updateUserAvatar(
   prisma: PrismaClient
 ) {
   try {
-    const { id } = req.params;
-    const userId = parseInt(id, 10);
+    const { userId } = req.params;
+    const userIdParsed = parseInt(userId, 10);
 
     if (!req.file) {
       return res.status(400).json({ error: 'No file uploaded' });
@@ -279,7 +289,7 @@ export async function updateUserAvatar(
       storageType === 'CLOUD' ? fileMetadata.cloudUrl : fileMetadata.savedPath;
 
     const updatedUser = await prisma.user.update({
-      where: { id: userId },
+      where: { id: userIdParsed },
       data: {
         avatarSource: fileSource,
       },
@@ -289,8 +299,10 @@ export async function updateUserAvatar(
       avatarSource: updatedUser.avatarSource,
       message: 'User avatar updated successfully',
     });
+    return;
   } catch (error) {
     console.error('Error updating user avatar: ', error);
     res.status(500).json({ error: 'Failed to update user avatar' });
+    return;
   }
 }
